Re-enable innodb
For some reason the database fails to bootstrap if innodb is disabled.  It might be that innodb can be disabled again after the database is created, not sure.  In the past i disabled innodb due to memory constraints on the server this was running on.  That is no longer the case, hence happy to keep it enabled even though I'm not using it